requirements.txt
# Core dependencies
torch>=1.10.0
transformers>=4.18.0
datasets>=2.0.0
numpy>=1.20.0
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0

# API and serving
fastapi>=0.70.0
uvicorn>=0.15.0
pydantic>=1.8.0
starlette>=0.17.0
python-multipart>=0.0.5
aiofiles>=0.8.0
httpx>=0.21.0
redis>=4.1.0

# Data processing
nltk>=3.6.0
spacy>=3.2.0
beautifulsoup4>=4.10.0
lxml>=4.6.0
PyYAML>=6.0.0
jsonschema>=4.4.0

# Biblical resources
bible-corpus>=1.0.0
pygreek>=1.0.0
pyhebrew>=1.0.0

# Text processing
regex>=2021.10.0
ftfy>=6.0.0
tqdm>=4.62.0
smart-open>=5.2.0

# Monitoring and logging
prometheus-client>=0.13.0
python-json-logger>=2.0.0
sentry-sdk>=1.5.0

# Testing
pytest>=6.2.0
pytest-cov>=3.0.0
pytest-mock>=3.6.0
requests-mock>=1.9.0

# Development tools
black>=22.1.0
flake8>=4.0.0
isort>=5.10.0
pre-commit>=2.17.0
mypy>=0.931
Sphinx>=4.4.0

# Security
pyjwt>=2.3.0
passlib>=1.7.0
cryptography>=36.0.0
python-dotenv>=0.19.0